refactor: Update get_page function to use string formatting for Redis key
This commit modifies the get_page function in web.py to use string formatting instead of concatenation when creating the Redis key for counting the number of requests made to a URL. By using string formatting, the code becomes more concise and easier to read. This change improves the maintainability of the codebase.